Add systemtz module with SystemTZ class
`SystemTZ` is a subclass of `datetime.tzinfo` and is API compatible with
`ZoneInfo`. Instances of the class represent the time zone the system
(or the application) is configured to.

It implements this via delegating to `time.*` functionality and is
sensitive to and supports changing the time zone via setting the `TZ`
environment variable and subsequently calling `time.tzset()` on systems
that provide this function (not Windows).

We also add a test suite with an extensive list of datetimes lots of
which are potentially tricky corner cases to ensure that the
implementation is not faulty.